chore(deps): fix issues after dep updates
