+++ packages/commitlint-config/tsconfig.json
@@ -1,9 +1,9 @@
 {
-	"extends": "../tsconfig/commonjs.json",
+	"extends": "../tsconfig/base.json",
 	"compilerOptions": {
+		"rootDir": "src",
 		"declaration": true,
-		"outDir": "dist",
-		"rootDir": "src"
+		"outDir": "dist"
 	},
 	"include": ["src/**/*"]
 }
