+++ packages/prettier-config/src/index.ts
@@ -5,11 +5,11 @@ const options: Config = {
 	endOfLine: 'lf',
 	printWidth: 100,
 	proseWrap: 'always',
-	singleQuote: true,
 	semi: true,
+	singleQuote: true,
 	trailingComma: 'all',
 	useTabs: true,
-
+	// eslint-disable-next-line perfectionist/sort-objects
 	overrides: [
 		{
 			files: '**/*.{yaml,yml}',
@@ -28,4 +28,4 @@ const options: Config = {
 	],
 };
 
-export = options;
+export default options;
